+++ test/components/underline_panel_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class PrimerUnderlinePanelTest < Minitest::Test
+  include Primer::ComponentTestHelpers
+
+  def test_renders
+    render_inline(Primer::UnderlinePanel.new)
+
+    assert_text("Add a test here")
+  end
+end